move model for export preparation to be able reuse in other exporters
